chore(): minor refactor
