+++ src/pages/details/positions-page/hooks/use-positions.ts
@@ -8,12 +8,12 @@ import {
   YieldBalancesWithIntegrationIdDto,
   useYieldFindValidators,
 } from "@stakekit/api-hooks";
-import { useState } from "react";
 import { Maybe } from "purify-ts";
 import { useDebounce } from "../../../../hooks/use-debounce";
 import { useTranslation } from "react-i18next";
 import { importValidator } from "../../../../common/import-validator";
 import { useSKWallet } from "../../../../providers/sk-wallet";
+import { useState } from "react";
 
 export const usePositions = () => {
   const { data, ...rest } = usePositionsData();
